+++ example/lib/main_appcast.dart
@@ -1,8 +1,9 @@
 /*
- * Copyright (c) 2019-2022 Larry Aasen. All rights reserved.
+ * Copyright (c) 2019-2024 Larry Aasen. All rights reserved.
  */
 
 import 'package:flutter/material.dart';
+import 'package:http/http.dart' as http;
 import 'package:upgrader/upgrader.dart';
 
 void main() async {
@@ -20,6 +21,8 @@ class MyApp extends StatelessWidget {
   static const appcastURL =
       'https://raw.githubusercontent.com/larryaasen/upgrader/master/test/testappcast.xml';
   final upgrader = Upgrader(
+    client: http.Client(),
+    clientHeaders: {'header1': 'value1'},
     storeController: UpgraderStoreController(
       onAndroid: () => UpgraderAppcastStore(appcastURL: appcastURL),
       oniOS: () => UpgraderAppcastStore(appcastURL: appcastURL),